requirements.txt
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
pandas==2.2.1
python-dateutil==2.9.0.post0
